test(torghut): split hyperliquid runtime service tests
